==> Task_2_Mann-Kendall/04 MK results comparison.R <==
rm(list=ls(all=TRUE)) 

# Load libraries
if(!require(tidyr)) { install.packages('tidyr') }; library(tidyr)
if(!require(plyr)) { install.packages('plyr') }; library(plyr)
if(!require(dplyr)) { install.packages('dplyr') }; library(dplyr)

# Set directories
dir.udf <- c('C:\\Users\\miles.medina\\Documents\\UF Postdoc\\CHNEP\\Task 2\\R files\\User-Defined Functions')
dir.dat <- c('C:\\Users\\miles.medina\\Documents\\UF Postdoc\\CHNEP\\Task 2\\R files\\Data')
dir.out <- c('C:\\Users\\miles.medina\\Documents\\UF Postdoc\\CHNEP\\Task 2\\R files\\Output')

# Load data 
setwd( dir.out )
  load("MK.trend.10y.USF.results.RData")  # 1st comparison, using current data requirements
  # load("MK.trend.10y.results.RData")  # 2nd comparison, using 'proposed' data requirements
setwd( dir.dat )
  USF.results <- read.csv('Trend Analysis - CHNEP - 2020 - Results.csv')
  grid.stations <- read.csv('grid-station-associations.csv')


# Format USF's trend analysis results
  # Subset 10-year trend results for 2011-2020 period
  USF.results <- USF.results[ which( USF.results$AnalysisTimeSpan=='10 year' ), ]
  USF.results <- USF.results[ which( USF.results$MinYear==2011 ), ]
  USF.results <- USF.results[ which( USF.results$MaxYear==2020 ), ]
  # Rename analytes ('Parameters') to match analyte names in CCS results
  analytes.old <- USF.results$Parameter %>% unique() %>% sort()
  analytes.new <- c( 'BOD', 'Chl-a', 'Color', 'Sp Cond', 'DO Conc', 'DO Sat', 'Fecal', 'NH4', 'NOx', 'PO4',
                     'pH', 'Temp', 'TKN', 'TN', 'TP', 'TSS', 'Turbidity' )
  USF.results$Analyte <- mapvalues( USF.results$Parameter, analytes.old, analytes.new ) 
    # QC check: Confirm analytes are matched correctly
    for( i in 1:length(analytes.old) ){
      cat( paste( analytes.old[i], "::", unique(USF.results$Analyte[which(USF.results$Parameter==analytes.old[i])]),"\n" ) )
    }  # // end i loop
  # Format unique station names consistent with the CCS results
  USF.results$DataSource <- sub( "WIN_", "", USF.results$DataSource ) # remove 'WIN_' prefixes
  USF.results$StationID <- sub( "GSV006         ", "GSV006", USF.results$StationID ) # remove extra spaces from GSV006
  USF.results$Unique.StationID <- paste0( USF.results$DataSource, ": ", USF.results$StationID )  # generate unique station IDs
  # Print Unique.StationID's appearing in USF.results but not appearing in CCS results
  missing.USF <- unique(USF.results$Unique.StationID)[ which( !( unique(USF.results$Unique.StationID) %in%
                                                                 unique(results.df$Station) ) ) ] %>% sort()
  missing.USF
  # Confirm that missing.USF stations are all associated with strata (and therefore should be missing from CCS results)
  match.stations <- logical()
  for( i in 1:length(missing.USF) ){
    this.source <- unique( USF.results$DataSource[ which( USF.results$Unique.StationID==missing.USF[i] ) ] )
    this.station <- unique( USF.results$StationID[ which( USF.results$Unique.StationID==missing.USF[i] ) ] )
    agencies <- data.frame( Name1 = c('SWFWMD','Lee County','Sarasota County','CHNEP'),
                            Name2 = c('SWFWMD','LEE_WQ','SARASOTAES_WQ','CHNEPMP') )
    match.idx <- which( grid.stations$Agency==agencies$Name1[which(agencies$Name2==this.source)] &
                        grid.stations$Station==this.station )
    # Print result to console
    if( length(match.idx==1) ){
      match.stations[i] <- TRUE
      cat( paste0("i=",i,". ",this.source,": ",this.station," matched ",grid.stations$Agency[match.idx],
                  " ",grid.stations$Station[match.idx],"\n") )
    } else if( length(match.idx)>1 ){
      match.stations[i] <- FALSE
      cat( paste0("i=",i,". ",this.source,": ",this.station,". Multiple matches found.\n") )
    } else {
      match.stations[i] <- FALSE
      cat( paste0("i=",i,". ",this.source,": ",this.station,". No match found.\n") )
    }
  }  # // end i loop
  all(match.stations)
  
  # Subset columns
  USF.results <- USF.results[ , which( colnames(USF.results) %in% c('Unique.StationID','Analyte','Trend') ) ]
  colnames( USF.results )[ which( colnames(USF.results)=='Unique.StationID' ) ] <- 'Station'
  # Coerce column classes
  USF.results$Station <- USF.results$Station %>% as.factor()
  USF.results$Trend <- USF.results$Trend %>% as.numeric()
  USF.results$Analyte <- USF.results$Analyte %>% as.factor()
  # Create new columns for MK trend (Trend negative, zero, positive) and for large trends (Trend of -2 or 2)
    # MK.trend (Positive Trends get labeled 1; negative Trends get labeled -1; non-sig Trends get labeled 0; others NA)
    USF.results$USF.MK.trend <- NA
    USF.results$USF.MK.trend[ which( USF.results$Trend>0 ) ] <- 1
    USF.results$USF.MK.trend[ which( USF.results$Trend==0 ) ] <- 0
    USF.results$USF.MK.trend[ which( USF.results$Trend<0 ) ] <- -1
    # trend.large (Trends of -1 or 1 get labeled FALSE; Trends of -2 or 2 get labeled TRUE; others NA)
    USF.results$USF.trend.large <- NA
    USF.results$USF.trend.large[ which( abs(USF.results$Trend)==1 ) ] <- FALSE
    USF.results$USF.trend.large[ which( abs(USF.results$Trend)==2 ) ] <- TRUE
  # Remove USF's original Trend column
    USF.results <- USF.results[, -which(colnames(USF.results)=='Trend') ]
    

    
# Compare USF and CCS trend results
    # Join CCS results (results.df) and USF results (USF.results)
    all.results <- full_join( x = results.df, y = USF.results, by = c('Station','Analyte') )
    # In a new column, label comparisons of results
    all.results$Compare <- NA
    all.results$Compare[ which( all.results$MK.trend != all.results$USF.MK.trend ) ] <- "MK.trend"
    all.results$Compare[ which( all.results$MK.trend == all.results$USF.MK.trend &
                                all.results$trend.large != all.results$USF.trend.large ) ] <- "trend.large"
    all.results$Compare[ which( ( all.results$MK.trend == all.results$USF.MK.trend |
                                 (is.na(all.results$MK.trend) & is.na(all.results$USF.MK.trend)) ) &
                                ( all.results$trend.large == all.results$USF.trend.large | 
                                 (is.na(all.results$trend.large) & is.na(all.results$USF.trend.large)) ) ) ] <- 'Match'
    all.results$Compare[ which( !is.na(all.results$MK.trend) & is.na(all.results$USF.MK.trend) ) ] <- 'CCSrun.USFnotrun'
    all.results$Compare[ which( is.na(all.results$MK.trend) & !is.na(all.results$USF.MK.trend) ) ] <- 'CCSnotrun.USFrun'
    # Summary of comparisons
    table( all.results$Compare, useNA='always' )
    # Stations labeled 'CCSnotrun.USFrun'
    all.results$Station[ which( all.results$Compare=="CCSnotrun.USFrun" ) ] %>% unique %>% sort
    # Stations labeled 'CCSrun.USFnotrun'
    all.results$Station[ which( all.results$Compare=="CCSrun.USFnotrun" ) ] %>% unique %>% sort
    
    
# Investigate disagreements in trend direction
    # USF trend up vs. CCS no trend
    all.results[ which( all.results$USF.MK.trend==1 &
                        all.results$MK.trend==0 ), ]
    # USF trend up vs. CCS trend down  
    all.results[ which( all.results$USF.MK.trend==1 &
                        all.results$MK.trend==-1 ), ] 
    # USF trend down vs. CCS no trend
    all.results[ which( all.results$USF.MK.trend==-1 &
                        all.results$MK.trend==0 ), ] 
    # USF trend down vs. CCS trend up  
    all.results[ which( all.results$USF.MK.trend==-1 &
                        all.results$MK.trend==1 ), ] 
    # USF no trend vs. CCS trend up
    all.results[ which( all.results$USF.MK.trend==0 &
                        all.results$MK.trend==1 ), ] 
    # USF no trend vs. CCS trend down
    all.results[ which( all.results$USF.MK.trend==0 &
                        all.results$MK.trend==-1 ), ] 
    # USF ran analysis vs. CCS did not run
    all.results[ which( !is.na(all.results$USF.MK.trend) &
                        is.na(all.results$MK.trend) ), ]
      # Unique stations run by USF but not run by CCS
      notrun.stations <- all.results$Station[ which( !is.na(all.results$USF.MK.trend) &
                                                     is.na(all.results$MK.trend) ) ] %>% unique() %>% sort()
        # Find stratum-associated stations in `notrun.stations`
        rm(match.idx)
        strata.stations <- logical()
        for( i in 1:length(notrun.stations) ){
          this.source <- strsplit( notrun.stations[i], ": " )[[1]][1]
          this.station <- strsplit( notrun.stations[i], ": " )[[1]][2]
          agencies <- data.frame( Name1 = c('SWFWMD','Lee County','Sarasota County','CHNEP'),
                                  Name2 = c('SWFWMD','LEE_WQ','SARASOTAES_WQ','CHNEPMP') )
          match.idx <- which( grid.stations$Agency==agencies$Name1[which(agencies$Name2==this.source)] &
                                grid.stations$Station==this.station )
          # Print result to console
          if( length(match.idx==1) ){
            strata.stations[i] <- TRUE
            cat( paste0("i=",i,". ",this.source,": ",this.station," matched ",grid.stations$Agency[match.idx],
                        " ",grid.stations$Station[match.idx],"\n") )
          } else if( length(match.idx)>1 ){
            strata.stations[i] <- FALSE
            cat( paste0("i=",i,". ",this.source,": ",this.station,". Multiple matches found.\n") )
          } else {
            strata.stations[i] <- FALSE
            cat( paste0("i=",i,". ",this.source,": ",this.station,". No match found.\n") )
          }
        }  # // end i loop
        # Print strata stations
        notrun.stations[which(strata.stations)]
        # Print non-strata stations
        notrun.stations[-which(strata.stations)]
        # Count 'not run' records associated with strata
        all.results[ which( !is.na(all.results$USF.MK.trend) &
                              is.na(all.results$MK.trend) &
                              all.results$Station %in% notrun.stations[which(strata.stations)] ), ] %>% nrow()
      
    # USF did not run vs. CCS ran analysis
    all.results[ which( is.na(all.results$USF.MK.trend) &
                        !is.na(all.results$MK.trend) ), ]
